create vote on createTrip page and renders correctly on specific trip page
