+++ client/src/components/VotingMech/index.js
@@ -9,43 +9,44 @@ const VotingMech = (props) => {
     const { currentTrip } = useContext(TripContext);
     const { currentUser } = useContext(UserContext);
     const [isAdmin, setAdmin] = useState([]);
-    const [currentVote, setCurrentVote] = useState([]);
+    const [currentVote, setCurrentVote] = useState({});
 
     useEffect(() => {
         // check if the user is an admin
         fetchCurrentVoteData();
         checkUserAdmin();
         // do i need something in here to fetch the vote data if the person isn't an admin?
-    }, [currentTrip, currentUser])
+    }, [currentTrip])
 
     function fetchCurrentVoteData() {
-        API.getSpecificTrip(currentTrip.id)
+        API.getSpecificTrip(props.tripId)
             .then(res => setCurrentVote(res.data))
     }
 
     function checkUserAdmin() {
         console.log(currentTrip.id);
         console.log(currentUser.id);
-        if(currentTrip.id && currentUser.id){
-        API.isUserAdmin({
-            user_id: currentUser.id,
-            trip_id: currentTrip.id
-        })
-            .then(res => setAdmin(res.data))
-            .catch(err => console.log(err))}
+        if (currentTrip.id && currentUser.id) {
+            API.isUserAdmin({
+                user_id: currentUser.id,
+                trip_id: currentTrip.id
+            })
+                .then(res => setAdmin(res.data))
+                .catch(err => console.log(err))
+        }
     }
 
     console.log(`this is the admin data ${isAdmin.admin}`)
 
-    function onCreateDate(data) {
-        API.setVote(
-            {
-                voteData: data,
-                trip: currentTrip.id
-            }
-        )
-            .then(res => console.log(res.data))
-    }
+    // function onCreateDate(data) {
+    //     API.setVote(
+    //         {
+    //             voteData: data,
+    //             trip: currentTrip.id
+    //         }
+    //     )
+    //         .then(res => console.log(res.data))
+    // }
 
     function onUpvoteDate(data, diff) {
         ;
@@ -79,15 +80,15 @@ const VotingMech = (props) => {
             .then(res => console.log(res.data))
     }
 
-    function onCreateLocation(data) {
-        API.setVote(
-            {
-                locationData: data,
-                trip: currentTrip.id
-            }
-        )
-            .then(res => console.log(res.data))
-    }
+    // function onCreateLocation(data) {
+    //     API.setVote(
+    //         {
+    //             locationData: data,
+    //             trip: currentTrip.id
+    //         }
+    //     )
+    //         .then(res => console.log(res.data))
+    // }
 
     function onUpvoteLocation(data, diff) {
         console.log(data)
@@ -120,15 +121,15 @@ const VotingMech = (props) => {
             .then(res => console.log(res.data))
     }
 
-    function onCreateActivity(data) {
-        API.setVote(
-            {
-                activityData: data,
-                trip: currentTrip.id
-            }
-        )
-            .then(res => console.log(res.data))
-    }
+    // function onCreateActivity(data) {
+    //     API.setVote(
+    //         {
+    //             activityData: data,
+    //             trip: currentTrip.id
+    //         }
+    //     )
+    //         .then(res => console.log(res.data))
+    // }
 
     function onUpvoteActivity(data, diff) {
         console.log(data)
@@ -161,15 +162,15 @@ const VotingMech = (props) => {
             .then(res => console.log(res.data))
     }
 
-    function onCreateTransport(data) {
-        API.setVote(
-            {
-                transportData: data,
-                trip: currentTrip.id
-            }
-        )
-            .then(res => console.log(res.data))
-    }
+    // function onCreateTransport(data) {
+    //     API.setVote(
+    //         {
+    //             transportData: data,
+    //             trip: currentTrip.id
+    //         }
+    //     )
+    //         .then(res => console.log(res.data))
+    // }
 
     function onUpvoteTransport(data, diff) {
         console.log(data)
@@ -202,73 +203,51 @@ const VotingMech = (props) => {
             .then(res => console.log(res.data))
     }
 
-    if (currentVote) {
         return (
             <div className="grid-container">
                 <div className="card" >
                     <h3>Dates:</h3>
                     <ReactVote data={currentVote.dateVote}
-                        onCreate={onCreateDate} 
-                        onUpvote={onUpvoteDate} 
-                        onClose={onCloseDate} 
-                        onReset={onResetDate} 
-                        isAdmin={isAdmin.admin} 
+                        // onCreate={onCreateDate}
+                        onUpvote={onUpvoteDate}
+                        onClose={onCloseDate}
+                        onReset={onResetDate}
+                        isAdmin={isAdmin.admin}
                         clientId={currentUser.id} />
                 </div>
                 <div className="card">
                     <h3>Locations:</h3>
                     <ReactVote data={currentVote.locationVote}
-                        onCreate={onCreateLocation} 
-                        onUpvote={onUpvoteLocation} 
-                        onClose={onCloseLocation} 
-                        onReset={onResetLocation} 
-                        isAdmin={isAdmin.admin} 
-                        clientId={currentUser.id}/>
+                        // onCreate={onCreateLocation}
+                        onUpvote={onUpvoteLocation}
+                        onClose={onCloseLocation}
+                        onReset={onResetLocation}
+                        isAdmin={isAdmin.admin}
+                        clientId={currentUser.id} />
                 </div>
                 <div className="card">
                     <h3>Activities:</h3>
                     <ReactVote data={currentVote.activityVote}
-                        onCreate={onCreateActivity} 
-                        onUpvote={onUpvoteActivity} 
-                        onClose={onCloseActivity} 
-                        onReset={onResetActivity} 
-                        isAdmin={isAdmin.admin} 
+                        // onCreate={onCreateActivity}
+                        onUpvote={onUpvoteActivity}
+                        onClose={onCloseActivity}
+                        onReset={onResetActivity}
+                        isAdmin={isAdmin.admin}
                         clientId={currentUser.id} />
                 </div>
                 <div className="card">
                     <h3>Mode of Transport:</h3>
                     <ReactVote data={currentVote.transportVote}
-                        onCreate={onCreateTransport} 
-                        onUpvote={onUpvoteTransport} 
-                        onClose={onCloseTransport} 
-                        onReset={onResetTransport} 
-                        isAdmin={isAdmin.admin} 
+                        // onCreate={onCreateTransport}
+                        onUpvote={onUpvoteTransport}
+                        onClose={onCloseTransport}
+                        onReset={onResetTransport}
+                        isAdmin={isAdmin.admin}
                         clientId={currentUser.id} />
                 </div>
             </div>
         )
-    } else {
-        return (
-            <div>
-                <div>
-                    <h3>Dates:</h3>
-                    <ReactVote onCreate={onCreateDate} onUpvote={onUpvoteDate} onClose={onCloseDate} onReset={onResetDate} isAdmin={isAdmin.admin} clientId={currentUser.id} />
-                </div>
-                <div>
-                    <h3>Locations:</h3>
-                    <ReactVote onCreate={onCreateLocation} onUpvote={onUpvoteLocation} onClose={onCloseLocation} onReset={onResetLocation} isAdmin={isAdmin.admin} clientId={currentUser.id} />
-                </div>
-                <div>
-                    <h3>Activities:</h3>
-                    <ReactVote onCreate={onCreateActivity} onUpvote={onUpvoteActivity} onClose={onCloseActivity} onReset={onResetActivity} isAdmin={isAdmin.admin} clientId={currentUser.id} />
-                </div>
-                <div>
-                    <h3>Mode of Transport:</h3>
-                    <ReactVote onCreate={onCreateTransport} onUpvote={onUpvoteTransport} onClose={onCloseTransport} onReset={onResetTransport} isAdmin={isAdmin.admin} clientId={currentUser.id} />
-                </div>
-            </div>
-        )
-    }
+    
 }
 
 export default VotingMech
