+++ client/src/pages/startNewTrip.js
@@ -4,6 +4,7 @@ import { Input, FormBtn } from "../components/Form";
 import API from "../utils/API";
 import { TripContext } from "../utils/TripContext";
 import { UserContext } from "../utils/UserContext";
+import ReactVote from 'react-vote';
 
 
 function StartNewTrip(props) {
@@ -11,6 +12,7 @@ function StartNewTrip(props) {
     const { currentTrip, setCurrentTrip } = useContext(TripContext);
 
     const [formObject, setFormObject] = useState({});
+    const [selectedTrip, setSelectedTrip] = useState({});
 
     function handleInputChange(event) {
         const { name, value } = event.target;
@@ -18,7 +20,7 @@ function StartNewTrip(props) {
     };
 
     function handleFormSubmit(event) {
-
+        alert(`The trip ${formObject.trip_name} was created!`)
         event.preventDefault();
         console.log(formObject.trip_name)
 
@@ -36,6 +38,46 @@ function StartNewTrip(props) {
             user_id: currentUser.id,
             admin: true
         })
+            .then(res => setSelectedTrip(res.data))
+    }
+
+    function onCreateDate(data) {
+        API.setVote(
+            {
+                voteData: data,
+                trip: selectedTrip.trip_id
+            }
+        )
+            .then(res => console.log(res.data))
+    }
+
+    function onCreateLocation(data) {
+        API.setVote(
+            {
+                locationData: data,
+                trip: selectedTrip.trip_id
+            }
+        )
+            .then(res => console.log(res.data))
+    }
+
+    function onCreateActivity(data) {
+        API.setVote(
+            {
+                activityData: data,
+                trip: selectedTrip.trip_id
+            }
+        )
+            .then(res => console.log(res.data))
+    }
+    
+    function onCreateTransport(data) {
+        API.setVote(
+            {
+                transportData: data,
+                trip: selectedTrip.trip_id
+            }
+        )
             .then(res => console.log(res.data))
     }
 
@@ -44,6 +86,7 @@ function StartNewTrip(props) {
         <div className="row">
             <div className="col">
                 <h2>Create New Trip</h2>
+                <h5>Step 1: Name your trip and click Create Trip</h5>
             </div>
             <div className="col">
                 <form>
@@ -59,6 +102,37 @@ function StartNewTrip(props) {
                         Create Trip
                     </button>
                 </form>
+                <div className="grid-container">
+                <h5>Step 2: Set your vote parameters for each category.</h5>
+                    <div className="card" >
+                        <h3>Dates:</h3>
+                        <ReactVote 
+                            onCreate={onCreateDate}
+                            isAdmin={true}
+                            clientId={currentUser.id} />
+                    </div>
+                    <div className="card">
+                        <h3>Locations:</h3>
+                        <ReactVote 
+                            onCreate={onCreateLocation}
+                            isAdmin={true}
+                            clientId={currentUser.id} />
+                    </div>
+                    <div className="card">
+                        <h3>Activities:</h3>
+                        <ReactVote
+                            onCreate={onCreateActivity}
+                            isAdmin={true}
+                            clientId={currentUser.id} />
+                    </div>
+                    <div className="card">
+                        <h3>Mode of Transport:</h3>
+                        <ReactVote 
+                            onCreate={onCreateTransport}
+                            isAdmin={true}
+                            clientId={currentUser.id} />
+                    </div>
+                </div>
             </div>
         </div>
     )
